requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
python-dotenv==1.0.1
pyserial==3.5
pexpect==4.9.0
PyYAML==6.0.1
gevent==24.2.1
gevent-websocket==0.10.1
